+++ CHANGELOG.md
@@ -6,7 +6,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-_In progress…_
+### Added
+- Support string casting for resources
 
 ## [1.2.1] - 2018-08-28
 ### Fixed